move/mobile(#18): web과 폴더 구조 통일
